Refactor parsing to regex only This removes the requirement for the lxml library
